organized files into a clean arch layer based package system
