+++ src/main/java/com/onionshop/entities/Brush.java
@@ -1,4 +1,4 @@
-package com.onionshop;
+package com.onionshop.entities;
 
 public abstract class Brush {
     private String brushShape;
