+++ src/main/java/module-info.java
@@ -5,9 +5,14 @@ module com.onionshop {
     requires org.controlsfx.controls;
     requires com.dlsc.formsfx;
     requires org.kordamp.bootstrapfx.core;
+    requires javafx.graphics;
 
     opens com.onionshop to javafx.fxml;
     opens com.onionshop.controllers to javafx.fxml;
 
     exports com.onionshop;
+    exports com.onionshop.managers;
+    opens com.onionshop.managers to javafx.fxml;
+    exports com.onionshop.entities;
+    opens com.onionshop.entities to javafx.fxml;
 }
